# Fix for white hairlines rendering in pdf