This fix will provide a solution to the following issues:

#3351
#4860
#3331
#9464 (Closed) (Duplicate)
#9481 (Closed) (Duplicate)
#9829 (Closed) (Duplicate)
#9953 (Closed) (Duplicate)

This solution takes into account the resolution of the viewer before painting images to the canvas and rounds the painting to the nearest full pixel. The problem in the issues above is caused by drawing to a canvas with a width/height that is more accurate than pixel width allows. This results in the drawing coloring part of a pixel, which causes a partial transparency. This transparency creates what appears to be a thin line between images. Rounding the images to the nearest pixel insures there will be no partially filled pixels, and therefore no thin lines.

Performance Tests
```
Run1:
browser | stat         | Count | Baseline(ms) | Current(ms) |  +/- |     %  | Result(P<.05)
------- | ------------ | ----- | ------------ | ----------- | ---- | ------ | -------------
Firefox | Overall      |    50 |        17427 |       16954 | -474 |  -2.72 |              
Firefox | Page Request |    50 |           11 |           9 |   -2 | -14.34 |              
Firefox | Rendering    |    50 |        17416 |       16944 | -472 |  -2.71 |   
```
```
Run2:
browser | stat         | Count | Baseline(ms) | Current(ms) |   +/- |     %  | Result(P<.05)
------- | ------------ | ----- | ------------ | ----------- | ----- | ------ | -------------
Firefox | Overall      |    50 |        18640 |       15158 | -3481 | -18.68 |        faster
Firefox | Page Request |    50 |           11 |          10 |    -1 |  -5.59 |              
Firefox | Rendering    |    50 |        18629 |       15148 | -3481 | -18.69 |        faster
```