@marktanner1331 2019-08-01
The issue is unfortunately more complex than believed:

A majority of PDF's will first set up a clipping path before drawing the image. The clipping path will be the exact dimensions of the image. 

This clipping path will suffer from the same sub-pixel rendering issues as described above, and cause the white lines to appear even though the image is technically the right size. 

To test this theory, try commenting out the clipping code in canvas.consumePath() leaving only the ctx.beginPath(); you'll notice that the white lines are gone.