@ssaring 2019-12-11
We have the same rendering problem in our current project, it's reported in issue #11400.
It would be very helpful for us to know whether there's a chance for solving these problems.

Unfortunately I have no programming skills in the PDF and rendering context, is there any other way to support the progress?