@jasonbolbar 2020-07-31
Is there any update for this? I'm interested in the solution for this issue. I tried to check the code cloning the repo from @brianholleibm and checking out to **3351_White_Bar_Fix** branch, later I ran **gulp generic** in order to generate the viewer and finally I tested in an iPhone using BrowserStack getting this result:
<img width="438" alt="phone_test" src="https://user-images.githubusercontent.com/6331505/89051442-7fc8eb00-d311-11ea-8bd8-515aa23dd927.png">