@Snuffleupagus 2021-08-19
Closing for now, since there's not been any updates for a couple of years and as-is this patch cannot land given that there's too many regressions in existing test-cases; a number of issues remain open to track this (please see https://github.com/mozilla/pdf.js/pull/10016#issue-211238996 for a list).